refactor(agent): split useAgentStream.ts around a shared flush-queue core
Tier 3 item from docs/reports/REPORT_LARGE_FILE_MODULARIZATION_SCAN_2026_07_22.md.
Crash-history-sensitive file — see RETRO_REPLACECHILD_CRASH_2026-06-06.md:
two independent documentAtom writes triggering separate reactive flushes in
the same browser task previously caused reconcileArrays/replaceChild
crashes. The single-RAF/single-batch() design is the deliberate fix, so the
split extracted the shared flush mechanism FIRST as stream-flush-queue.ts
(the ONLY requestAnimationFrame and ONLY batch() call site for the whole
hook, independently grep-verified — every match elsewhere is a warning
comment, not a second call site), then pulled each event-source producer
(useToolChunkStream, useShellNodeStream, useTurnLifecycle,
usePendingMessageAcceptance) out as hooks that push into that one shared
queue instance rather than scheduling their own flush. Every new file
carries an explicit comment warning against ever adding a second RAF/batch()
site.

useAgentStream.ts: 951 -> 483 lines (orchestrator only: NDJSON parse loop,
translator/parser state, composes the four extracted hooks).

tsc --noEmit clean; the whole frontend/app/view/agent test tree (61 files,
809 tests) passes unchanged. Independently re-verified the single-RAF/
single-batch invariant via grep before committing.

<!-- agentmux:agent_id=agentx -->
